refactor(cli): split MCP tools registry by domain (issue #70)
Move tool_specs() into per-domain files under mcp/tools/ (core, interact,
inspect, eval, observe, storage, assert, record). Each file exposes
fn specs() -> Vec<ToolSpec>; tools() in mod.rs concatenates them.

Schema helpers split between mcp/schemas.rs (generic) and
mcp/tools/schemas.rs (domain-specific). Kills
#[allow(clippy::too_many_lines)] on the old monolithic tool_specs().
Allow count: 12 → 11.
